Module pmcore::data

New data format

Modules

pmetrics

Structs

Bolus

An instantaenous input of drug

Covariate

A Covariate is a collection of CovariateSegments, which allows for interpolation of covariate values

CovariateSegment

A CovariateSegment is a segment of the piece-wise interpolation of a Covariate

Covariates

Covariates is a collection of Covariate

Data

Data is a collection of Subjects, which are collections of Occasions, which are collections of Events

Infusion

A continuous dose of drug

Observation

An observation of drug concentration or covariates

Occasion

An Occasion is a collection of events, for a given Subject, that are from a specific occasion

Subject

Subject is a collection of blocks for one individual

Enums

Event

An Event can be a Bolus, Infusion, or Observation

InterpolationMethod

Traits

CovariateInterpolator

CovariateTrait

CovariatesTrait

DataTrait

OccasionTrait

Each Subject can have multiple occasions

SubjectTrait

Subject is a trait that represents a single individual in a dataset