Function pmcore::entrypoints::simulate

pub fn simulate(
    _equation: Equation,
    _data: Data,
    _settings: Settings,
) -> Result<()>

Simulate predictions from a model and prior distribution

This function is used to simulate predictions from a model and prior distribution. The output is a CSV file with the following columns:

id: subject ID, corresponding to the desired dose regimen

point: support point index (0-indexed)

time: prediction time

pred: simulated prediction

Arguments

The user can specify the desired settings in a TOML configuration file, see routines::settings::simulator for details.

idelta: the interval between predictions. Default is 0.0.

tad: the time after dose, which if greater than the last prediction time is the time for which it will predict . Default is 0.0.