pmcore
Module prelude

A collection of commonly used items to simplify imports.

Re-exports

pub use crate::algorithms;

pub use crate::algorithms::dispatch_algorithm;

pub use crate::routines;

pub use crate::routines::logger;

pub use crate::routines::logger::setup_log;

pub use crate::routines::settings;

Modules

builder

covariate

data

error_model

event

models

prelude

simulator

structs

Macros

dmatrix

Construct a dynamic matrix directly from data.

fa

fetch_cov

fetch_params

lag

Structs

Analytical

Bolus

An instantaenous input of drug

Covariate

A Covariate is a collection of CovariateSegments, which allows for interpolation of covariate values

CovariateSegment

A CovariateSegment is a segment of the piece-wise interpolation of a Covariate

Covariates

Covariates is a collection of Covariate

Data

Data is a collection of Subjects, which are collections of Occasions, which are collections of Events

ErrorModel

HashMap

A hash map implemented with quadratic probing and SIMD lookup.

Infusion

A continuous dose of drug

Meta

This structs holds the metadata of the model

ODE

Observation

An observation of drug concentration or covariates

Occasion

An Occasion is a collection of events, for a given Subject, that are from a specific occasion

SDE

Subject

Subject is a collection of blocks for one individual

Enums

ErrorType

Event

An Event can be a Bolus, Infusion, or Observation

InterpolationMethod

Traits

Equation

EquationTypes

EstimateTheta

OptimalSupportPoint

Predictions

State

SubjectBuilderExt

Functions

one_compartment

Analytical for one compartment Assumptions:

one_compartment_with_absorption

Analytical for one compartment with absorption Assumptions:

two_compartments

Analytical for two compartment Assumptions:

two_compartments_with_absorption

Analytical for two compartment with absorption Assumptions:

Type Aliases

Result

Result<T, Error>