Trait pmcore::routines::evaluation::sigma::Sigma

pub trait Sigma {
    // Required method
    fn sigma(&self, yobs: &Array1<f64>) -> Array1<f64>;
}

Contains information on the observation error

Required Methods

fn sigma(&self, yobs: &Array1<f64>) -> Array1<f64>

Estimates the standard deviation of the observation error for given observations.

Arguments

yobs - A 1-dimensional Array containing observed values.

Returns

A 1-dimensional Array representing the estimated standard deviation of the observation error.