Function pmcore::routines::initialization::sobol::generate

pub fn generate(
    n_points: usize,
    range_params: &Vec<(f64, f64)>,
    seed: usize
) -> ArrayBase<OwnedRepr<f64>, Dim<[usize; 2]>>

Generates a 2-dimensional array containing a Sobol sequence within the given ranges

Returns

A 2D array where each row is a point, and each column corresponds to a parameter.