Module pmcore::routines::settings

Provides routines for reading and parsing settings

Structs

Advanced

This struct contains advanced options and hyperparameters

Config

General configuration settings

Constant

Parameters which are held constant

Convergence

This struct contains the convergence criteria for the algorithm

Error

Defines the error model and polynomial to be used

Fixed

Parameters which are estimated, but fixed for the population

Log

Output

Configuration for the output files

Predictions

Prior

Configuration for the prior

Random

Random parameters to be estimated

Settings

Contains all settings for PMcore

Functions

read

Parses the settings from a TOML configuration file

write_settings_to_file

Writes a copy of the parsed settings to file