pmcore::routines::settings
Function read

pub fn read(path: impl Into<String>) -> Result<Settings, Error>

Parses the settings from a TOML configuration file

This function parses the settings from a TOML configuration file. The settings are validated, and a copy of the settings is written to file.

Entries in the TOML file may be overridden by environment variables. The environment variables must be prefixed with PMCORE_, and the TOML entry must be in uppercase. For example, the TUI may be disabled by setting the environment variable PMCORE_CONFIG_TUI=false A single underscore, _, is used as the separator for nested entries.