Module pmcore::routines::simulation::predict

Structs

Engine

Model

Traits

Predict

return the predicted values for the given scenario and parameters where the second element of the tuple is the predicted values one per observation time in scenario and in the same order it is not relevant the outeq of the specific event.

Functions

get_ypred

post_predictions

sim_obs

Simulate observations for multiple scenarios and support points.

simple_sim