Function pmcore::routines::simulation::predict::get_population_predictions

pub fn get_population_predictions<'a>(
    equation: &'a Equation,
    subjects: &Vec<Subject>,
    support_points: &Array2<f64>,
    _cache: bool
) -> PopulationPredictions

Number of support points to cache for each scenario return the predicted values for the given scenario and parameters where the second element of the tuple is the predicted values one per observation time in scenario and in the same order it is not relevant the outeq of the specific event. Simulate observations for multiple scenarios and support points.

This function performs simulation of observations for multiple scenarios and support points using the provided simulation engine sim_eng. It returns a 2D Array where each element represents the simulated observations for a specific scenario and support point.

Arguments

sim_eng - A reference to the simulation engine implementing the Predict trait.

scenarios - A reference to a Vec<Scenario> containing information about different scenarios.

support_points - A 2D Array (Array2<f64>) representing the support points. Each row corresponds to a different support point scenario.

cache - A boolean flag indicating whether to cache predicted values during simulation.

Returns

A 2D Array (Array2<Array1<f64>>) where each element is an Array1<f64> representing the simulated observations for a specific scenario and support point.

Example

In this example, observations will contain the simulated observations for multiple scenarios and support points.

Note: This function allows for optional caching of predicted values, which can improve performance when simulating observations for multiple scenarios.