Trait pmcore::prelude::Equation

pub trait Equation: EquationPriv + 'static + Clone + Sync {
    // Required method
    fn estimate_likelihood(
        &self,
        subject: &Subject,
        support_point: &Vec<f64>,
        error_model: &ErrorModel<'_>,
        cache: bool,
    ) -> f64;

    // Provided methods
    fn estimate_predictions(
        &self,
        subject: &Subject,
        support_point: &Vec<f64>,
    ) -> Self::P { ... }
    fn simulate_subject(
        &self,
        subject: &Subject,
        support_point: &Vec<f64>,
        error_model: Option<&ErrorModel<'_>>,
    ) -> (Self::P, Option<f64>) { ... }
}

Required Methods

fn estimate_likelihood( &self, subject: &Subject, support_point: &Vec<f64>, error_model: &ErrorModel<'_>, cache: bool, ) -> f64

Estimate the likelihood of the subject given the support point and error model This function might be cached

Provided Methods

fn estimate_predictions( &self, subject: &Subject, support_point: &Vec<f64>, ) -> Self::P

fn simulate_subject( &self, subject: &Subject, support_point: &Vec<f64>, error_model: Option<&ErrorModel<'_>>, ) -> (Self::P, Option<f64>)

Object Safety

This trait is not object safe.