Module pmcore::routines::output

Routines for writing results to file, such as predicted values

Structs

CycleLog

This holdes a vector of NPCycle objects to provide a more detailed log

NPCycle

An NPCycle object contains the summary of a cycle It holds the following information:

NPResult

Defines the result objects from an NPAG run An NPResult contains the necessary information to generate predictions and summary statistics

OutputFile

Contains all the necessary information of an output file

Functions

median

population_mean_median

posterior

posterior_mean_median

write_pmetrics_observations