Module pmcore::routines::settings

Provides routines for reading and parsing settings

Structs

Config

General configuration settings

Constant

Parameters which are held constant

Error

Defines the error model and polynomial to be used

Fixed

Parameters which are estimated, but fixed for the population

Paths

This struct contains the paths to the data, log and prior files.

Random

Random parameters to be estimated

Settings

Contains all settings for PMcore

Functions

read_settings

Parses the settings from a TOML configuration file

write_settings_to_file

Writes a copy of the parsed settings to file