pmcore::routines
Module settings

Structs

Advanced

This struct contains advanced options and hyperparameters

AlgorithmSet

Config

General configuration settings

Convergence

This struct contains the convergence criteria for the algorithm

Error

Defines the error model and polynomial to be used

ErrorSet

InitialState

Log

Output

Configuration for the output files

Parameter

Defines a parameter to be estimated

Parameters

This structure contains information on all Parameters to be estimated

ParametersBuilder

Builder for creating a set of parameters

ParametersSet

Predictions

Prior

Configuration for the prior

Settings

Contains all settings for PMcore

SettingsBuilder

Enums

LogLevel

The log level, which can be one of the following:

Traits

AlgorithmDefined

ErrorModelDefined

ParametersDefined

Functions

read

Parses the settings from a TOML configuration file